changed AUTH to TS
The AUTH file is now in type script usage is 'let auth = new Authentication("token","userid");'  we will need to pull the user id from firebot I don't know if its exposed. WE should have a toke file that is git Ignored because this needs to be a private key found in the extension setup page.